requirements.txt
Flask
python-pptx
requests
beautifulsoup4
PyMuPDF
pdfplumber
matplotlib
Pillow
graphviz